refactor the builder API to avoid symbol table manipulation
the builder API was using symbol tables internally, which meant that
datalog elements were converted too early, and tat made the API too
complicate. This aligns the builder implementation with the approach use
in the rust implementation, where the builder types only contain other
builder types, and the conversion with the symbol table is only done
when calling the build() method.
This also removes some functions from the public API that could be
misused to get invalid symbol tables